requirements.txt
opencv-python-headless
Pillow
networkx
pandas
python-louvain
matplotlib
